created ticker price component
